requirements.txt
Django==1.10.1
django-tinymce==2.4.0
